reusable-downloadUtils.ts-made
